http: unembed http.Server from HTTPServer
Initially this change is being made to support registering endpoints from a separate
package. In order to ensure none of the endpoint methods use this field it needs
to be unembedded, so the IDE can find usages.

In the future HTTPServer may be able to remove the field entirely, and can become a
handler which is passed to an http.Server